requirements.txt
requests
Mastodon.py
python-dotenv
lxml
tldextract
sqlalchemy